refactor(schemas): relocate streamData parsing schemas to core
Move the @agent-free per-stream sidecar parsing schemas + legacy-flatten
helpers from the extension's streamTabSchemas.ts into host-agnostic
src/shared/schemas/streamData.ts so the upcoming core StreamSnapshotStore and
the extension managers decode identical on-disk shapes. streamTabSchemas.ts
keeps only StreamTabMetaSchema (needs @agent TaskState) + re-exports.

Phase 2 prep.
