+++ src/shared/schemas/index.ts
@@ -35,6 +35,7 @@ export * from './todo';
 export * from './todoDisplay';
 export * from './plan';
 export * from './workPlan';
+export * from './streamData';
 export * from './subagentProgress';
 export * from './inquiry';
 export * from './prompts';
